+++ docs/blocks/index.ts
@@ -13,3 +13,4 @@ export * from "./ComponentList";
 export * from "./FontPreview";
 export * from "./FontTokens";
 export * from "./Table";
+export * from "./Callout";