+++ src/components/Form/components/FormText.component.tsx
@@ -1,16 +1,16 @@
 import { Text } from "@/components/Text";
-import { GridItem } from "@layouts";
 
 import type { BuiltFormTextProps } from "../Form.types";
+import { FormFieldGridItem } from "./FormFieldGridItem.component";
 
 export const FormText = (props: BuiltFormTextProps) => {
   const { kind: _, children, ...textProps } = props;
 
   return (
-    <GridItem columnStart={1} columnEnd={-1} fill>
+    <FormFieldGridItem columnStart={1} columnEnd={-1} fill>
       <Text {...textProps} testId="form-text" overrideTestId>
         {children}
       </Text>
-    </GridItem>
+    </FormFieldGridItem>
   );
 };