+++ src/components/Hypertext/Hypertext.component.tsx
@@ -16,7 +16,7 @@ const defaultHypertextProps: DefaultHypertextProps = {
 /**
  * Renders a hyperlink component with a text and an icon.
  *
- * @version 0.0.1
+ * @version 0.0.2
  *
  * @param {HypertextProps} props - The hypertext's props
  * @return {JSX.Element} The rendered hypertext
@@ -45,7 +45,7 @@ const Hypertext = (props: HypertextProps): JSX.Element => {
       rel="noreferrer"
       data-testid="hypertext"
     >
-      <Text size="m" weight="medium" color="primary-base">
+      <Text size="m" weight="medium" color="primary-base" underline>
         {children}
       </Text>
       <Icon name="BoxArrowUpRight" size="s-3" color="primary-base" />